+++ libraries/TM1637_RT/TM1637.h
@@ -3,17 +3,18 @@
 //    FILE: TM1637.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2019-10-28
-// VERSION: 0.3.6
+// VERSION: 0.3.7
 // PUPROSE: TM1637 library for Arduino
 //     URL: https://github.com/RobTillaart/TM1637_RT
 
 
-// tested on 6 digit display + 4 digit (clock) display
+//  tested on 6 digit display
+//  tested on 4 digit (clock) display esp. displayTime()
 
 
 #include "Arduino.h"
 
-#define TM1637_LIB_VERSION      (F("0.3.6"))
+#define TM1637_LIB_VERSION      (F("0.3.7"))
 
 
 class TM1637
@@ -33,19 +34,24 @@ public:
   void displayFloat(float value, uint8_t fixedPoint);
   void displayHex(uint32_t value);
   void displayClear();
-  //  only works on 4 digit display with colon
+
+  //  next 3 only tested on 4 digit display with colon
   void displayTime(uint8_t hh, uint8_t mm, bool colon);
+  void displayTwoInt(int ll, int rr, bool colon = true);
+  void displayCelsius(int temp, bool colon = false);
 
 
   //  BRIGHTNESS
-  void    setBrightness(uint8_t brightness);
+  //  brightness = 0..7
+  void    setBrightness(uint8_t brightness = 3);  //  default 3
   uint8_t getBrightness();
 
 
   //  BIT DELAY
   //  tune the timing of writing bytes.
-  void    setBitDelay(uint8_t bitDelay = 10) { _bitDelay = bitDelay; };
-  uint8_t getBitDelay() { return _bitDelay; };
+  void    setBitDelay(uint8_t bitDelay = 10);
+  uint8_t getBitDelay();
+
 
   //  KEY SCAN
   uint8_t keyscan(void);
@@ -53,6 +59,7 @@ public:
 
   //  CONFIGURATION
   //  the order the individual digits must be sent to the display.
+  //  optionally to be called after begin()
   void    setDigitOrder(uint8_t a = 0, uint8_t b = 1,
                         uint8_t c = 2, uint8_t d = 3,
                         uint8_t e = 4, uint8_t f = 5,
@@ -67,7 +74,7 @@ public:
 private:
   uint8_t _clock      = -1;
   uint8_t _data       = -1;
-  uint8_t _digits     = 4;
+  uint8_t _digits     = 6;
   uint8_t _brightness = 3;
   uint8_t _bitDelay   = 10;